v1.0

+++ CodeTxFreeRtos/USER/TASK/game_task.c
@@ -5,7 +5,7 @@
 
 void game_task(void *pvParameters)
 {
-	uint16_t game = pvParameters;
+	uint16_t game = (uint16_t)pvParameters;
 	if(game == 1)
 		ball_game_run();
 	else if(game == 2)
